wip start cleaning
