+++ front/lib/actions/mcp_internal_actions/output_schemas.ts
@@ -633,6 +633,21 @@ const RenderedNodeSchema = z.object({
   connectorProvider: z.enum(CONNECTOR_PROVIDERS).nullable(),
 });
 
+const RenderedWarehouseNodeSchema = z.object({
+  nodeId: z.string(),
+  title: z.string(),
+  path: z.string(),
+  parentTitle: z.string().nullable(),
+  mimeType: z.string(),
+  hasChildren: z.boolean(),
+  connectorProvider: z.enum(CONNECTOR_PROVIDERS).nullable(),
+  sourceUrl: z.undefined(),
+  lastUpdatedAt: z.undefined(),
+});
+export type RenderedWarehouseNodeType = z.infer<
+  typeof RenderedWarehouseNodeSchema
+>;
+
 export const DataSourceNodeListSchema = z.object({
   mimeType: z.literal(INTERNAL_MIME_TYPES.TOOL_OUTPUT.DATA_SOURCE_NODE_LIST),
   uri: z.literal(""),
@@ -654,18 +669,14 @@ export const isDataSourceNodeListType = (
   );
 };
 
-// Schema for tables filesystem browsing (no query field)
 export const TablesFilesystemBrowseSchema = z.object({
   mimeType: z.literal(
     "application/vnd.dust.tool-output.tables-filesystem-browse"
   ),
   uri: z.literal(""),
   text: z.string(),
-  nodeId: z
-    .string()
-    .nullable()
-    .describe("The node ID that was browsed (null for root)"),
-  data: z.array(RenderedNodeSchema),
+  nodeId: z.string().nullable(),
+  data: z.array(RenderedWarehouseNodeSchema),
   nextPageCursor: z.string().nullable(),
   resultCount: z.number(),
 });
